scripts/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
email-validator==2.1.0
pydantic[email]==2.5.0